requirements.txt
requests>=2.28.0
Pillow>=9.0.0
cairosvg>=2.5.0